fixed last statetype

+++ inc/state_types.h
@@ -9,7 +9,7 @@ typedef struct {
 	uint32_t cell_max_mV;
 	uint32_t cell_capacity_cAh;
 	uint32_t num_modules;
-	uint16_t cell_charge_c_rating_cC;
+	uint32_t cell_charge_c_rating_cC;
 	uint32_t bal_on_thresh_mV;
 	uint32_t bal_off_thresh_mV;
 	uint32_t pack_cells_p;
